Moved convolution functions into convolution.py
